pyproject.toml
[project]
name = "account"
version = "0.1.0"
description = "AVWX account management backend service"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Michael duPont", email="michael@dupont.dev"}
]
dependencies = [
    "bcrypt==4.1.2",
    "beanie==1.25.0",
    "fastapi==0.110.0",
    "fastapi-jwt==0.2.0",
    "fastapi-mail==1.4.1",
    "httpx==0.27.0",
    "kewkew==0.1",
    "logfire[fastapi,system-metrics,pymongo]==0.32.1",
    "mailchimp3==3.0.21",
    "python-decouple==3.8",
    "rollbar==1.0.0",
    "stripe==8.8.0",
    "uvicorn==0.29.0",
]

[project.urls]
Issues = "https://github.com/avwx-rest/account-backend/issues"
Source = "https://github.com/avwx-rest/account-backend"

[tool.hatch.envs.types]
extra-dependencies = [
    "mypy>=1.0.0"
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:account tests}"

[tool.hatch.envs.hatch-test]
extra-dependencies = [
    "asgi-lifespan>=2.1",
    "pytest-asyncio>=0.23",
    "pytest-cov",
]

[tool.hatch.envs.serve.scripts]
main = "uvicorn account.main:app --port {args:8080}"
reload = "uvicorn account.main:app --reload --port {args:8080}"

[tool.pytest.ini_options]
addopts = """\
    --cov account \
    --cov tests \
    --cov-report term-missing \
    --no-cov-on-fail \
"""

[tool.coverage.run]
source_pkgs = ["account", "tests"]
branch = true
parallel = true

[tool.coverage.report]
fail_under = 65  # increase over time
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
check_untyped_defs = true
disallow_any_unimported = false
disallow_untyped_defs = true
explicit_package_bases = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
lint.ignore = [
    "D105",
    "D203",
    "D213",
    "B008", # Depends in func def
    "T201",
    "INP001",
]